Adds redux to user sign up log in
